Initial Implementation of Genesis Codes Using JNI Interface
This is the initial implementation of a demonstration Genesis Codes
using the Java JNI interface. The encoding and decoding is in (16,3).
There will be 16 data shards and 3 parity shards for each strip of
data. The data shards are in uints of 384 bytes and parity shards
are in units of 400 bytes in length.

To build:
1. At the root level, type: gradle build
2. Go to the build/ folder, initialize cmake once: cmake ..
3. Still under the build folder, issue: make (or make -j4)

Note step 2 and 3 does not apply in this open source project.
You must contact Anthony Mai, inventor of Genesis Codes, to
obtain the JNI component libgenesis.so for an evaluation.

To test:
Under the build/ folder, issue this java command:
java -Djava.library.path=. -cp ./classes/java/main com/genesis/GenesisEncoder foo.txt